+++ NEWS.md
@@ -1,5 +1,8 @@
 # fs (development version)
 
+* `path_package()` now works with paths in development packages automatically
+  (#175).
+
 # fs 1.2.7
 
 ## New features